requirements.txt
python-dotenv
fastapi
uvicorn
pandas
numpy
matplotlib
scipy
seaborn
xlrd
openpyxl
lxml
requests
beautifulsoup4
